backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
boto3==1.34.79
botocore==1.34.79
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.4
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.15.1
exceptiongroup==1.2.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
jmespath==1.0.1
oauthlib==3.2.2
openai==1.16.2
packaging==24.0
pillow==10.3.0
psycopg==3.1.18
psycopg-binary==3.1.18
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
